requirements.txt
APScheduler==3.8.0
asgiref==3.4.1
astroid==2.8.0
colorama==0.4.4
Django==3.2.7
django-cors-headers==3.8.0
django-dramatiq==0.10.0
djangorestframework==3.12.4
dramatiq==1.11.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pika==1.2.0
platformdirs==2.3.0
plumbum==1.7.0
prometheus-client==0.11.0
psycopg2-binary==2.9.1
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
redis==3.5.3
rpyc==5.0.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
wrapt==1.12.1
